[compile] Migrate Types to a dedicated module
stack-info: PR: https://github.com/pytorch/helion/pull/2603, branch: hinriksnaer/stack/12
